+++ packages/frontend/src/components/table/TableComponent.tsx
@@ -17,6 +17,8 @@ import LastPageIcon from "@mui/icons-material/LastPage";
 import { TableHead } from "@mui/material";
 import MyDialog from "@components/MyDiaLog";
 import axios, { AxiosResponse } from "axios";
+import { TableColumms } from '@components/table/TableColumns';
+import { TableItems } from '@components/table/TableItems'
 
 interface TablePaginationActionsProps {
   count: number;
@@ -108,34 +110,7 @@ function TablePaginationActions(props: TablePaginationActionsProps) {
   );
 }
 
-function createData(
-  staffid: number,
-  firstname: string,
-  lastname: string,
-  zoneid: number,
-  branchid: number,
-  telno: string
-) {
-  return { staffid, firstname, lastname, zoneid, branchid, telno };
-}
-
-const rows = [
-  createData(1, "Eeieiza", "hahaluvluv", 2, 3, "xxxx-xxx-xxx"),
-  createData(2, "Eeieiza", "hahaluvluv", 2, 3, "xxxx-xxx-xxx"),
-  createData(3, "Eeieiza", "hahaluvluv", 2, 3, "xxxx-xxx-xxx"),
-  createData(4, "Eeieiza", "hahaluvluv", 2, 3, "xxxx-xxx-xxx"),
-  createData(5, "Eeieiza", "hahaluvluv", 2, 3, "xxxx-xxx-xxx"),
-  createData(6, "Eeieiza", "hahaluvluv", 2, 3, "xxxx-xxx-xxx"),
-  createData(7, "Eeieiza", "hahaluvluv", 2, 3, "xxxx-xxx-xxx"),
-  createData(8, "Eeieiza", "hahaluvluv", 2, 3, "xxxx-xxx-xxx"),
-  createData(9, "Eeieiza", "hahaluvluv", 2, 3, "xxxx-xxx-xxx"),
-  createData(10, "Eeieiza", "hahaluvluv", 2, 3, "xxxx-xxx-xxx"),
-  createData(11, "Eeieiza", "hahaluvluv", 2, 3, "xxxx-xxx-xxx"),
-  createData(12, "Eeieiza", "hahaluvluv", 2, 3, "xxxx-xxx-xxx"),
-  createData(13, "Eeieiza", "hahaluvluv", 2, 3, "xxxx-xxx-xxx"),
-].sort((a, b) => (a.staffid < b.staffid ? -1 : 1));
-
-export function TableTest() {
+export const TableComponent = ({ title, columns, rows, style }) => {
   const [page, setPage] = React.useState(0);
   const [rowsPerPage, setRowsPerPage] = React.useState(5);
 
@@ -161,45 +136,14 @@ export function TableTest() {
     <div className="h-full">
       <div>
         <TableContainer component={Paper} className="mx-auto ">
+          {title}
           <Table aria-label="custom pagination table">
-            <TableHead>
-              <TableRow>
-                <TableCell>StaffId</TableCell>
-                <TableCell>FirstName</TableCell>
-                <TableCell>LastName</TableCell>
-                <TableCell>ZoneId</TableCell>
-                <TableCell>BranchId</TableCell>
-                <TableCell>Tel-no</TableCell>
-              </TableRow>
-            </TableHead>
+            <TableColumms names={columns} />
             <TableBody>
               {(rowsPerPage > 0
-                ? rows.slice(
-                    page * rowsPerPage,
-                    page * rowsPerPage + rowsPerPage
-                  )
-                : rows
-              ).map((row) => (
-                <TableRow key={row.staffid}>
-                  <TableCell style={{ width: 160 }}>{row.staffid}</TableCell>
-                  <TableCell style={{ width: 160 }}>{row.firstname}</TableCell>
-                  <TableCell style={{ width: 160 }}>{row.lastname}</TableCell>
-                  <TableCell style={{ width: 160 }}>{row.zoneid}</TableCell>
-                  <TableCell style={{ width: 160 }}>{row.branchid}</TableCell>
-                  <TableCell style={{ width: 160 }}>{row.telno}</TableCell>
-                  <div className="space-x-3">
-                    <button className="p-2 font-semibold bg-green-400 rounded-md">
-                      Edit
-                    </button>
-                    <button className="p-2 font-semibold bg-red-400 rounded-md">
-                      Delete
-                    </button>
-                    <div>
-                      <MyDialog />
-                    </div>
-                  </div>
-                </TableRow>
-              ))}{" "}
+                ? <TableItems rows={rows.slice(page * rowsPerPage, page * rowsPerPage + rowsPerPage)}/> 
+                : <TableItems rows={rows} />
+              )}
               {emptyRows > 0 && (
                 <TableRow style={{ height: 53 * emptyRows }}>
                   <TableCell colSpan={6} />
